Version 1.4.0.

+++ vine/updateView.py
@@ -6,6 +6,7 @@ import option
 import utility
 import grapeGit as git
 import grapeConfig
+import grapeMenu
 import checkout
 
 
@@ -13,7 +14,8 @@ import checkout
 class UpdateView(option.Option):
     """
     grape uv  - Updates your active submodules and ensures you are on a consistent branch throughout your project.
-    Usage: grape-uv [-f ] [-v] [--checkSubprojects] [-b]
+    Usage: grape-uv [-f ] [-v] [--checkSubprojects] [-b] [--skipSubmodules] [--allSubmodules]
+                    [--skipNestedSubprojects] [--allNestedSubprojects]
 
     Options:
         
@@ -24,6 +26,8 @@ class UpdateView(option.Option):
                                 not go through the 'which submodules do you want' script.
         -b                      Automatically creates subproject branches that should be there according to your branching
                                 model. 
+        --allSubmodules         Automatically add all submodules to your workspace. 
+        --allNestedSubprojects  Automatically add all nested subprojects to your workspace. 
 
     """
     def __init__(self):
@@ -79,7 +83,7 @@ class UpdateView(option.Option):
                     included[subproject] = False
             if opt.lower()[0] == "s":
                 for subproject in subprojects: 
-                    included[subproject] = utility.userInput("Would you like %s %s? [y/n]" % (projectType, subprojects),
+                    included[subproject] = utility.userInput("Would you like %s %s? [y/n]" % (projectType, subproject),
                                                             'n')
         for subprojects in toplevelSubs:
             included[subprojects] = utility.userInput("Would you like %s %s? [y/n]" % (projectType, subprojects), 'n')
@@ -96,14 +100,19 @@ class UpdateView(option.Option):
     def execute(self, args):
         config = grapeConfig.grapeConfig()
         quiet = not args["-v"]
+        origwd = os.getcwd()
+        os.chdir(utility.workspaceDir())
         base = git.baseDir()
         if base == "":
             return False
-        hasSubmodules = len(git.getAllSubmodules()) > 0
+        hasSubmodules = len(git.getAllSubmodules()) > 0 and not args["--skipSubmodules"]
         if not args["--checkSubprojects"]:
             # handle submodules first
             if hasSubmodules:
-                includedSubmodules = self.defineActiveSubmodules(quiet=quiet)
+                if args["--allSubmodules"]: 
+                    includedSubmodules = {sub:True for sub in git.getAllSubmodules(quiet = quiet)}
+                else:
+                    includedSubmodules = self.defineActiveSubmodules(quiet=quiet)
                 initStr = ""
                 if args["-f"]:
                     deinitStr = "-f"
@@ -116,9 +125,8 @@ class UpdateView(option.Option):
                         deinitStr += ' %s' % submodule
 
                 utility.printMsg("Configuring submodules...")
-                git.submodule("init", quiet=quiet)
-                os.chdir(git.baseDir())
                 utility.printMsg("Initializing submodules...")
+                git.submodule("init %s" % initStr.strip(), quiet=quiet)
                 if deinitStr or deinitStr == "-f":
                     utility.printMsg("Deiniting submodules that were not requested... (%s)" % deinitStr)
                     git.submodule("deinit %s" % deinitStr.strip(), quiet=quiet)
@@ -128,44 +136,51 @@ class UpdateView(option.Option):
                     git.submodule("update", quiet=quiet)
 
             # handle nested subprojects
-            os.chdir(base)
-            includedNestedSubprojectPrefixes = self.defineActiveNestedSubprojects(quiet=quiet)
-
-            allNestedSubprojects = config.getAllNestedSubprojects()
-            reverseLookupByPrefix = {config.get("nested-%s" % sub, "prefix") : sub for sub in allNestedSubprojects} 
-
-            userConfig = grapeConfig.grapeUserConfig()
-            updatedActiveList = []
-            for subproject, nowActive in includedNestedSubprojectPrefixes.items():
-                section = "nested-%s" % reverseLookupByPrefix[subproject]
-                userConfig.ensureSection(section)
-                previouslyActive = userConfig.getboolean(section, "active")
-
-                if nowActive and previouslyActive:
-                    updatedActiveList.append(subproject)
-
-                if nowActive and not previouslyActive:
-                    utility.printMsg("Activating Nested Subproject %s" % subproject)
+            if not args["--skipNestedSubprojects"]: 
+                
+                nestedPrefixLookup = lambda x : config.get("nested-%s" % x, "prefix")
+                allNestedSubprojects = config.getAllNestedSubprojects()
+                if args["--allNestedSubprojects"]: 
+                    includedNestedSubprojectPrefixes = {nestedPrefixLookup(sub):True for sub in allNestedSubprojects}
+                else:
+                    includedNestedSubprojectPrefixes = self.defineActiveNestedSubprojects(quiet=quiet)
+                reverseLookupByPrefix = {nestedPrefixLookup(sub) : sub for sub in allNestedSubprojects} 
+                userConfig = grapeConfig.grapeUserConfig()
+                updatedActiveList = []
+                for subproject, nowActive in includedNestedSubprojectPrefixes.items():
                     subprojectName = reverseLookupByPrefix[subproject]
-                    addSubproject.AddSubproject.activateNestedSubproject(subprojectName, userConfig)
-                    updatedActiveList.append(subprojectName)
-
-                if not nowActive and not previouslyActive:
-                    pass
-                if not nowActive and previouslyActive:
-                    #remove the submodule
-                    subprojectdir = os.path.join(base, utility.makePathPortable(subproject))
-                    proceed = args["-f"] or \
-                              utility.userInput("About to delete all contents in %s. Any uncommitted changes, branches "
-                                                "that are not pushed, or ignored files will be removed.  Proceed?" %
-                                                subproject, 'n')
-                    if proceed:
-                        shutil.rmtree(subprojectdir)
-            userConfig.setActiveNestedSubprojects(updatedActiveList)
-            grapeConfig.writeConfig(userConfig, os.path.join(utility.workspaceDir(), ".grapeuserconfig"))
+                    section = "nested-%s" % reverseLookupByPrefix[subproject]
+                    userConfig.ensureSection(section)
+                    previouslyActive = userConfig.getboolean(section, "active")
+                    previouslyActive = previouslyActive and os.path.exists(os.path.join(base, subproject))
+                    userConfig.set(section, "active", "True" if previouslyActive else "False")
+                    if nowActive and previouslyActive:
+                        updatedActiveList.append(subprojectName)
+    
+                    if nowActive and not previouslyActive:
+                        utility.printMsg("Activating Nested Subproject %s" % subproject)
+                        addSubproject.AddSubproject.activateNestedSubproject(subprojectName, userConfig)
+                        updatedActiveList.append(subprojectName)
+    
+                    if not nowActive and not previouslyActive:
+                        pass
+                    if not nowActive and previouslyActive:
+                        #remove the subproject
+                        subprojectdir = os.path.join(base, utility.makePathPortable(subproject))
+                        proceed = args["-f"] or \
+                                  utility.userInput("About to delete all contents in %s. Any uncommitted changes, committed changes "
+                                                    "that have not been pushed, or ignored files will be lost.  Proceed?" %
+                                                    subproject, 'n')
+                        if proceed:
+                            shutil.rmtree(subprojectdir)
+                userConfig.setActiveNestedSubprojects(updatedActiveList)
+                grapeConfig.writeConfig(userConfig, os.path.join(utility.workspaceDir(), ".git", ".grapeuserconfig"))
 
         checkoutArgs = "-b" if args["-b"] else ""
 
+        if args["--checkSubprojects"]:
+            utility.printMsg("Making sure all submodules are initialized...")
+            git.submodule("init")
         for subproject in grapeConfig.GrapeConfigParser.getAllActiveNestedSubprojectPrefixes():
             #ensure nested subprojects are on the appropriate branch (nested projects should have same branch layout)
             # as outer level repo. 
@@ -175,7 +190,7 @@ class UpdateView(option.Option):
             self.safeSwitchHeadlessRepoToBranch(subproject, desiredSubprojectBranch, checkoutArgs, quiet)
 
 
-        # ensure submodule is on apppropriate branch
+        # ensure submodule is on appropriate branch
         if config.getboolean("workspace", "manageSubmodules"):
             desiredSubmoduleBranch = self.getDesiredSubmoduleBranch(config)
             activeSubmodules = git.getActiveSubmodules(quiet=quiet)
@@ -184,6 +199,8 @@ class UpdateView(option.Option):
             for sub in activeSubmodules:
                 utility.printMsg("Ensuring %s is on %s" % (sub, desiredSubmoduleBranch))
                 self.safeSwitchHeadlessRepoToBranch(sub, desiredSubmoduleBranch, checkoutArgs, quiet)
+
+        os.chdir(origwd)
 
         return True
 
@@ -209,7 +226,21 @@ class UpdateView(option.Option):
             return
 
         if git.hasBranch(branch):
-            git.fetch("origin", "%s:%s" % (branch, branch), quiet=quiet)
+            try:
+                git.fetch("origin", "%s:%s" % (branch, branch), quiet=quiet)
+            except git.GrapeGitError as e:
+                if "[rejected]" in e.gitOutput and "(non-fast-forward)" in e.gitOutput:
+                    utility.printMsg("Fetch of %s rejected as non-fast-forward\nAttempting push of local %s in %s" % (branch, branch, repo))
+                    try:
+                        git.push("origin %s" % branch, quiet=quiet)
+                    except git.GrapeGitError as e2:
+                        utility.printMsg("Local and remote versions of %s may have diverged in %s" % (branch, repo))
+                        utility.printMsg("%s" % e2.gitOutput)
+                        mr = utility.userInput("Would you like to attempt to merge the remote using grape mr [y/n]", 'n')
+                        if mr:
+                            grapeMenu.menu().applyMenuChoice("mr", ["mr", branch])
+                else:    
+                    raise(e)
 
         checkout.Checkout.handledCheckout(checkoutArgs, branch, repo, quiet=quiet)
